refactor: split command logic into subs
